chore(escrow): fix baseline compile errors
- Remove duplicate get_version method (the one referencing non-existent DataKey::SchemaVersion)
- Rename symbol_short!("fund_st_ch") to "fstate_ch" globally (Soroban 9-char limit; 13 test sites)
- Rename symbol_short!("al_prm_upd") to "alprm_upd"
- Rename persistent_ttl_min_extension_ledgers (36 chars over 30 limit) to persistent_ttl_min_ext_ledgers (29 chars)
- Fix duplicate enum discriminant AllowlistParametersOutOfRange=72 (collision with TargetNotPositive=72) to 302
- Move escrow/src/tests.rs (232 lines) to escrow/src/tests/mod.rs (canonical Rust 2018 layout)
- Fix typoed EscrowError variants in collateral_boundary_tests.rs

Refs: liquifact-contracts baseline build, prerequisite for #1072 and #1038
