+++ escrow/src/tests/collateral_boundary_tests.rs
@@ -219,7 +219,7 @@ fn test_sme_commitment_zero_amount_rejected() {
     // Record at 0 (should be rejected)
     assert_contract_error(
         client.try_record_sme_collateral_commitment(&asset, &0i128),
-        EscrowError::InvalidCollateralAmount,
+        EscrowError::CollateralAmountNotPositive,
     );
 
     // Commitment should remain None
@@ -238,7 +238,7 @@ fn test_sme_commitment_negative_amount_rejected() {
     // Record at -1 (should be rejected)
     assert_contract_error(
         client.try_record_sme_collateral_commitment(&asset, &-1i128),
-        EscrowError::InvalidCollateralAmount,
+        EscrowError::CollateralAmountNotPositive,
     );
 
     // Commitment should remain None
@@ -257,7 +257,7 @@ fn test_sme_commitment_empty_asset_symbol_rejected() {
     // Record with empty asset symbol (should be rejected)
     assert_contract_error(
         client.try_record_sme_collateral_commitment(&asset, &1_000i128),
-        EscrowError::InvalidAssetSymbol,
+        EscrowError::CollateralAssetEmpty,
     );
 
     // Commitment should remain None
